working on bug fixes

+++ app/app/src/main/java/com/stockxsteals/app/viewmodel/db/PremiumViewModel.kt
@@ -23,9 +23,9 @@ class PremiumViewModel (
      }
    }
 
-   suspend fun getIsPremium(id: Int): Int {
+   suspend fun getIsPremium(id: Long): Int {
      return withContext(Dispatchers.IO) {
-       premiumDataSource.getIsPremium(id.toLong())
+       premiumDataSource.getIsPremium(id)
      }
    }
